Feature/nos tale update

See merge request GameField/NosWings_Old/Emulator!16

+++ srcs/OpenNos.Data/Interfaces/IItemInstance.cs
@@ -20,7 +20,7 @@ namespace OpenNos.Data.Interfaces
     {
         #region Properties
 
-        byte Amount { get; set; }
+        ushort Amount { get; set; }
 
         long? BoundCharacterId { get; set; }
